moving around old functions + initial phf_pbp function
